fix:ディレクトリ構成統一
