+++ src/app/analytics/page.tsx
@@ -1,7 +1,4 @@
-import { BarGraph } from '@/components/Chart/bar';
-import { LineGraph } from '@/components/Chart/line';
-import { View } from '@/components/Chart/view';
-import { Title } from '@/components/Title';
+import AnalyticsPage from '@/components/pages/analytics';
 
 export function generateMetadata() {
   return {
@@ -10,21 +7,5 @@ export function generateMetadata() {
 }
 
 export default function Analytics(): JSX.Element {
-  return (
-    <div className='flex flex-1 flex-col gap-4 pt-0'>
-      <Title text='アクセス解析' />
-      <p>Google AnalyticsからAPIでデータを取得し、ここで表示しています。</p>
-      <div className='grid auto-rows-min gap-4 md:grid-cols-3'>
-        <div className='rounded-xl bg-muted/50'>
-          <BarGraph />
-        </div>
-        <div className='rounded-xl bg-muted/50'>
-          <View />
-        </div>
-        <div className='rounded-xl bg-muted/50'>
-          <LineGraph />
-        </div>
-      </div>
-    </div>
-  );
+  return <AnalyticsPage />;
 }
